Use distinct path renderer type for each zip mover type
