+++ src/Entities/ConnectedStuff/ConnectedZipMover.cs
@@ -1,5 +1,4 @@
-﻿using Celeste.Mod.CommunalHelper.Entities;
-using Celeste.Mod.Entities;
+﻿using Celeste.Mod.Entities;
 using Microsoft.Xna.Framework;
 using Monocle;
 using System;
@@ -8,7 +7,169 @@ using System.Collections.Generic;
 
 namespace Celeste.Mod.CommunalHelper {
     [CustomEntity("CommunalHelper/ConnectedZipMover")]
-    public class ConnectedZipMover : ConnectedSolid, IMultiNodeZipMover {
+    public class ConnectedZipMover : ConnectedSolid {
+        public class PathRenderer : Entity {
+            private class Segment {
+                public bool Seen { get; set; }
+
+                private readonly Vector2 from, to;
+                private readonly Vector2 dir, twodir, perp;
+                private float length;
+
+                private readonly Vector2 lineStartA, lineStartB;
+                private readonly Vector2 lineEndA, lineEndB;
+
+                public Rectangle Bounds { get; }
+
+                private readonly Vector2 sparkAdd;
+
+                private readonly float sparkDirStartA, sparkDirStartB;
+                private readonly float sparkDirEndA, sparkDirEndB;
+
+                const float piOverEight = MathHelper.PiOver4 / 2f;
+                const float eightPi = 4 * MathHelper.TwoPi;
+
+                public Segment(Vector2 from, Vector2 to) {
+                    this.from = from;
+                    this.to = to;
+
+                    dir = (to - from).SafeNormalize();
+                    twodir = 2 * dir;
+                    perp = dir.Perpendicular();
+                    length = Vector2.Distance(from, to);
+
+                    Vector2 threeperp = 3 * perp;
+                    Vector2 minusfourperp = -4 * perp;
+
+                    lineStartA = from + threeperp;
+                    lineStartB = from + minusfourperp;
+                    lineEndA = to + threeperp;
+                    lineEndB = to + minusfourperp;
+
+                    sparkAdd = (from - to).SafeNormalize(5f).Perpendicular();
+                    float angle = (from - to).Angle();
+                    sparkDirStartA = angle + piOverEight;
+                    sparkDirStartB = angle - piOverEight;
+                    sparkDirEndA = angle + MathHelper.Pi - piOverEight;
+                    sparkDirEndB = angle + MathHelper.Pi + piOverEight;
+
+                    Rectangle b = Util.Rectangle(from, to);
+                    b.Inflate(10, 10);
+
+                    Bounds = b;
+                }
+
+                public void Spark(Level level) {
+                    level.ParticlesBG.Emit(ZipMover.P_Sparks, from + sparkAdd + Calc.Random.Range(-Vector2.One, Vector2.One), sparkDirStartA);
+                    level.ParticlesBG.Emit(ZipMover.P_Sparks, from - sparkAdd + Calc.Random.Range(-Vector2.One, Vector2.One), sparkDirStartB);
+                    level.ParticlesBG.Emit(ZipMover.P_Sparks, to + sparkAdd + Calc.Random.Range(-Vector2.One, Vector2.One), sparkDirEndA);
+                    level.ParticlesBG.Emit(ZipMover.P_Sparks, to - sparkAdd + Calc.Random.Range(-Vector2.One, Vector2.One), sparkDirEndB);
+                }
+
+                public void Render(float percent, Color rope, Color lightRope) {
+                    Draw.Line(lineStartA, lineEndA, rope);
+                    Draw.Line(lineStartB, lineEndB, rope);
+
+                    for (float d = 4f - percent * eightPi % 4f; d < length; d += 4f) {
+                        Vector2 teethA = lineStartA + perp + dir * d;
+                        Vector2 teethB = lineEndB - dir * d;
+                        Draw.Line(teethA, teethA + twodir, lightRope);
+                        Draw.Line(teethB, teethB - twodir, lightRope);
+                    }
+                }
+
+                public void RenderShadow(float percent) {
+                    Vector2 startA = lineStartA + Vector2.UnitY;
+                    Vector2 endB = lineEndB + Vector2.UnitY;
+
+                    Draw.Line(startA, lineEndA + Vector2.UnitY, Color.Black);
+                    Draw.Line(lineStartB + Vector2.UnitY, endB, Color.Black);
+
+                    for (float d = 4f - percent * eightPi % 4f; d < length; d += 4f) {
+                        Vector2 teethA = startA + perp + dir * d;
+                        Vector2 teethB = endB - dir * d;
+                        Draw.Line(teethA, teethA + twodir, Color.Black);
+                        Draw.Line(teethB, teethB - twodir, Color.Black);
+                    }
+                }
+            }
+
+            private readonly Rectangle bounds;
+            private readonly Segment[] segments;
+
+            private readonly ConnectedZipMover zipMover;
+
+            private Level level;
+
+            private readonly Color color, lightColor;
+
+            private readonly Vector2[] nodes;
+
+            public PathRenderer(ConnectedZipMover zipMover, Vector2[] nodes, Color color, Color lightColor) {
+                this.zipMover = zipMover;
+
+                this.nodes = new Vector2[nodes.Length];
+
+                Vector2 offset = new(zipMover.MasterWidth / 2f, zipMover.MasterHeight / 2f);
+
+                Vector2 prev = this.nodes[0] = nodes[0] + offset;
+                Vector2 min = prev, max = prev;
+
+                segments = new Segment[nodes.Length - 1];
+                for (int i = 0; i < segments.Length; ++i) {
+                    Vector2 node = this.nodes[i + 1] = nodes[i + 1] + offset;
+                    segments[i] = new(node, prev);
+
+                    min = Util.Min(min, node);
+                    max = Util.Max(max, node);
+
+                    prev = node;
+                }
+
+                bounds = new((int) min.X, (int) min.Y, (int) (max.X - min.X), (int) (max.Y - min.Y));
+                bounds.Inflate(10, 10);
+
+                this.color = color;
+                this.lightColor = lightColor;
+
+                Depth = Depths.SolidsBelow;
+            }
+
+            public override void Added(Scene scene) {
+                base.Added(scene);
+                level = scene as Level;
+            }
+
+            public void CreateSparks() {
+                foreach (Segment seg in segments)
+                    seg.Spark(level);
+            }
+
+            public override void Render() {
+                Rectangle cameraBounds = level.Camera.GetBounds();
+
+                if (!cameraBounds.Intersects(bounds))
+                    return;
+
+                foreach (Segment seg in segments)
+                    if (seg.Seen = cameraBounds.Intersects(seg.Bounds))
+                        seg.RenderShadow(zipMover.Percent);
+
+                foreach (Segment seg in segments)
+                    if (seg.Seen)
+                        seg.Render(zipMover.Percent, color, lightColor);
+
+                float rotation = zipMover.Percent * MathHelper.TwoPi;
+                foreach (Vector2 node in nodes) {
+                    zipMover.cog.DrawCentered(node + Vector2.UnitY, Color.Black, 1f, rotation);
+                    zipMover.cog.DrawCentered(node, Color.White, 1f, rotation);
+                }
+
+                zipMover.DrawBorder();
+            }
+        }
+        private PathRenderer zipMoverPathRenderer;
+
         public enum Themes {
             Normal,
             Moon,
@@ -29,9 +190,6 @@ namespace Celeste.Mod.CommunalHelper {
 
         // Lightning.
         private BloomPoint bloom;
-
-        // The instance of the PathRenderer Class defined above.
-        private ZipMoverPathRenderer zipMoverPathRenderer;
 
         private bool permanent;
         private bool waits;
@@ -188,7 +346,7 @@ namespace Celeste.Mod.CommunalHelper {
             base.Added(scene);
 
             // Creating the Path Renderer.
-            scene.Add(zipMoverPathRenderer = new(this, MasterWidth, MasterHeight, nodes, ropeColor, ropeLightColor, depth: Depths.SolidsBelow));
+            scene.Add(zipMoverPathRenderer = new(this, nodes, ropeColor, ropeLightColor));
         }
 
         public override void Removed(Scene scene) {
@@ -289,11 +447,6 @@ namespace Celeste.Mod.CommunalHelper {
                         (int) extension.Width + 2,
                         (int) extension.Height + 2),
                         Color.Black);
-        }
-
-        public void DrawCog(Vector2 node, float rotation) {
-            cog.DrawCentered(node + Vector2.UnitY, Color.Black, 1f, rotation);
-            cog.DrawCentered(node, Color.White, 1f, rotation);
         }
 
         private IEnumerator Sequence() {
